requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
pydantic==2.7.1
requests==2.32.3
python-dotenv==1.0.1
pydantic-settings==2.2.1